requirements.txt
streamlit 
groq 
chromadb 
sentence-transformers 
python-dotenv 
pandas 
